Fixed tests
Tests now look for key words in the stdout of tools to signify that the
help menu was reached succesfully.

Dummy “make test” targets added to tachyon, spark, and ADAM containers.

STAR has no help menu, and it seemed stupid to make “FATAL ERROR” a
success menu, so it has been replaced with a dummy “make test”.