Adds containers necessary for running Hadoop 2.6.2 HDFS.
This commit adds the containers necessary for running version 2.6.2 of the
Hadoop Distributed File System (HDFS). This includes:

* apache-hadoop-common: A container that contains setup which is common to both
  the worker and master nodes.
* apache-hadoop-master: A container that sets up a HDFS NameNode.
* apache-hadoop-worker: A container that sets up a HDFS DataNode.

The README.md files for each of these containers (as well as the
apache-spark-master, apache-spark-worker, and adam containers) describe the
way that the containers should be run to configure a HDFS + Spark cluster
where ADAM can run. These containers have been verified on a single node,
local cluster. The containers we have built could conceptually be used to
run Hadoop 2.6.2 YARN, but they do not currently set up or configure YARN.