Modified adam-pipeline to pass default memory (resolves #151)
I've modified the wrapper script for the adam-pipeline image to pass the memory
requested by the user as the default memory to request for a job. This resolves
an issue where a user running on a machine with a low amount of memory gets OOM
asserts from Toil, even though the Spark job doesn't request much memory.