Log failed tools in jenkins.py and exit if a build failed (resolves #211)
Adds logic to the `run_make` step in jenkins.py that tracks all build stages
that have failed. At the end, if one or more command failed, then all failed
commands are printed, and we exit with a non-zero exit code, which will cause
Jenkins to mark the build as failed. As opposed to throwing an exception whenever
a command fails, this approach allows us to run through all of the builds, which
is preferable from a debugging perspective.